[General] Tracking to upstream, rev 2015-01-14.

+++ source/src/emu.cpp
@@ -306,19 +306,22 @@ void EMU::notify_power_off()
 
 _TCHAR* EMU::bios_path(_TCHAR* file_name)
 {
-#if defined(_USE_AGAR) || defined(_USE_SDL) || defined(_USE_QT)
-        static _TCHAR file_path[_MAX_PATH];
-        strcpy(file_path, app_path);
-        strcat(file_path, file_name);
-#if defined(_USE_AGAR) || defined(_USE_SDL) || defined(_USE_QT)
-        AGAR_DebugLog(AGAR_LOG_INFO, "LOAD BIOS: %s\n", file_path);
-#endif
-#else
-        static _TCHAR file_path[_MAX_PATH];
-	_stprintf(file_path, _T("%s%s"), app_path, file_name);
-        printf("LOAD: %s\n", file_path);
-#endif
-	return file_path;
+ 	static _TCHAR file_path[_MAX_PATH];
+	_stprintf_s(file_path, _MAX_PATH, _T("%s%s"), app_path, file_name);
+ 	return file_path;
+//#if defined(_USE_AGAR) || defined(_USE_SDL) || defined(_USE_QT)
+//        static _TCHAR file_path[_MAX_PATH];
+//        strcpy(file_path, app_path);
+//        strcat(file_path, file_name);
+//#if defined(_USE_AGAR) || defined(_USE_SDL) || defined(_USE_QT)
+//        AGAR_DebugLog(AGAR_LOG_INFO, "LOAD BIOS: %s\n", file_path);
+//#endif
+//#else
+//        static _TCHAR file_path[_MAX_PATH];
+//	_stprintf(file_path, _T("%s%s"), app_path, file_name);
+//        printf("LOAD: %s\n", file_path);
+//#endif
+//	return file_path;
 }
 
 void EMU::suspend()
@@ -403,7 +406,7 @@ void EMU::open_printer_file()
 {
 	cur_time_t time;
 	get_host_time(&time);
-	_stprintf(prn_file_name, _T("prn_%d-%0.2d-%0.2d_%0.2d-%0.2d-%0.2d.txt"), time.year, time.month, time.day, time.hour, time.minute, time.second);
+	_stprintf_s(prn_file_name, _MAX_PATH, _T("prn_%d-%0.2d-%0.2d_%0.2d-%0.2d-%0.2d.txt"), time.year, time.month, time.day, time.hour, time.minute, time.second);
 	prn_fio->Fopen(bios_path(prn_file_name), FILEIO_WRITE_BINARY);
 }
 
@@ -453,7 +456,9 @@ void EMU::printer_strobe(bool value)
 #ifdef _DEBUG_LOG
 void EMU::initialize_debug_log()
 {
-	debug_log = _tfopen(_T("d:\\debug.log"), _T("w"));
+	if(_tfopen_s(&debug_log, _T("d:\\debug.log"), _T("w")) != 0) {
+		debug_log = NULL;
+	}
 }
 
 void EMU::release_debug_log()
@@ -472,13 +477,13 @@ void EMU::out_debug_log(const _TCHAR* format, ...)
 	static _TCHAR prev_buffer[1024] = {0};
 	
 	va_start(ap, format);
-	_vstprintf(buffer, format, ap);
+	_vstprintf_s(buffer, 1024, format, ap);
 	va_end(ap);
 	
 	if(_tcscmp(prev_buffer, buffer) == 0) {
 		return;
 	}
-	_tcscpy(prev_buffer, buffer);
+	_tcscpy_s(prev_buffer, 1024, buffer);
 	
 	if(debug_log) {
 		_ftprintf(debug_log, _T("%s"), buffer);
@@ -486,9 +491,11 @@ void EMU::out_debug_log(const _TCHAR* format, ...)
 		if((size += _tcslen(buffer)) > 0x8000000) { // 128MB
 			static int index = 1;
 			TCHAR path[_MAX_PATH];
-			_stprintf(path, _T("d:\\debug_#%d.log"), ++index);
+			_stprintf_s(path, _MAX_PATH, _T("d:\\debug_#%d.log"), ++index);
 			fclose(debug_log);
-			debug_log = _tfopen(path, _T("w"));
+			if(_tfopen_s(&debug_log, path, _T("w")) != 0) {
+				debug_log = NULL;
+			}
 			size = 0;
 		}
 	}
@@ -499,7 +506,7 @@ void EMU::out_message(const _TCHAR* format, ...)
 {
 	va_list ap;
 	va_start(ap, format);
-	_vstprintf(message, format, ap);
+	_vstprintf_s(message, 1024, format, ap);
 	va_end(ap);
 	message_count = 4; // 4sec
 }
@@ -621,7 +628,7 @@ void EMU::open_cart(int drv, _TCHAR* file_path)
 {
 	if(drv < MAX_CART) {
 		vm->open_cart(drv, file_path);
-		_tcscpy(cart_status[drv].path, file_path);
+		_tcscpy_s(cart_status[drv].path, _MAX_PATH, file_path);
 		out_message(_T("Cart%d: %s"), drv + 1, file_path);
 		
 		// restart recording
@@ -674,7 +681,7 @@ void EMU::open_disk(int drv, _TCHAR* file_path, int offset)
 			vm->open_disk(drv, file_path, offset);
 			out_message(_T("FD%d: %s"), drv + FD_BASE_NUMBER, file_path);
 		}
-		_tcscpy(disk_status[drv].path, file_path);
+		_tcscpy_s(disk_status[drv].path, _MAX_PATH, file_path);
 		disk_status[drv].offset = offset;
 	}
 }
@@ -727,7 +734,7 @@ void EMU::open_quickdisk(int drv, _TCHAR* file_path)
 			vm->open_quickdisk(drv, file_path);
 			out_message(_T("QD%d: %s"), drv + QD_BASE_NUMBER, file_path);
 		}
-		_tcscpy(quickdisk_status[drv].path, file_path);
+		_tcscpy_s(quickdisk_status[drv].path, _MAX_PATH, file_path);
 	}
 }
 
@@ -766,7 +773,7 @@ void EMU::play_tape(_TCHAR* file_path)
 		vm->play_tape(file_path);
 		out_message(_T("CMT: %s"), file_path);
 	}
-	_tcscpy(tape_status.path, file_path);
+	_tcscpy_s(tape_status.path, _MAX_PATH, file_path);
 	tape_status.play = true;
 }
 
@@ -785,7 +792,7 @@ void EMU::rec_tape(_TCHAR* file_path)
 		vm->rec_tape(file_path);
 		out_message(_T("CMT: %s"), file_path);
 	}
-	_tcscpy(tape_status.path, file_path);
+	_tcscpy_s(tape_status.path, _MAX_PATH, file_path);
 	tape_status.play = false;
 }
 
@@ -818,7 +825,7 @@ void EMU::open_laser_disc(_TCHAR* file_path)
 		vm->open_laser_disc(file_path);
 		out_message(_T("LD: %s"), file_path);
 	}
-	_tcscpy(laser_disc_status.path, file_path);
+	_tcscpy_s(laser_disc_status.path, _MAX_PATH, file_path);
 }
 
 void EMU::close_laser_disc()
@@ -884,19 +891,23 @@ void EMU::update_config()
 void EMU::save_state()
 {
 	_TCHAR file_name[_MAX_PATH];
-	_stprintf(file_name, _T("%s.sta"), _T(CONFIG_NAME));
+	_stprintf_s(file_name, _MAX_PATH, _T("%s.sta"), _T(CONFIG_NAME));
 	save_state_tmp(bios_path(file_name));
 }
 
 void EMU::load_state()
 {
 	_TCHAR file_name[_MAX_PATH];
-	_stprintf(file_name, _T("%s.sta"), _T(CONFIG_NAME));
-	save_state_tmp(bios_path(_T("$temp$.sta")));
-	if(!load_state_tmp(bios_path(file_name))) {
-		load_state_tmp(bios_path(_T("$temp$.sta")));
+	_stprintf_s(file_name, _MAX_PATH, _T("%s.sta"), _T(CONFIG_NAME));
+        FILEIO ffp;
+	if(ffp.IsFileExists(bios_path(file_name))) {
+		save_state_tmp(bios_path(_T("$temp$.sta")));
+		if(!load_state_tmp(bios_path(file_name))) {
+			out_debug_log("failed to load state file\n");
+			load_state_tmp(bios_path(_T("$temp$.sta")));
+		}
+		DeleteFile(bios_path(_T("$temp$.sta")));
 	}
-	DeleteFile(bios_path(_T("$temp$.sta")));
 }
 
 void EMU::save_state_tmp(_TCHAR* file_path)
